% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\docType{data}
\name{distribution}
\alias{distribution}
\title{Distribution of four taxa.}
\format{
A matrix object with two areas (rows) and four species (columns).
}
\usage{
distribution
}
\description{
A matrix with four species distributed in two areas.
}
\author{
Miranda-Esquivel Daniel R.
}
\keyword{datasets}
