% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalTerminalvsInternal.R
\name{evalTerminalvsInternal}
\alias{evalTerminalvsInternal}
\title{evalTerminalvsInternal}
\usage{
evalTerminalvsInternal(
  tree = tree,
  distribution = distribution,
  root = TRUE,
  valueAssign = 1,
  branch = "terminals",
  model = "allswap",
  nTimes = 100
)
}
\arguments{
\item{tree}{is a single tree with T terminals, an APER phylo object.}

\item{distribution}{is a labeled matrix object, with the distribution of T terminals (rows) in A areas (columns).}

\item{root}{is boolean (default=TRUE).}

\item{valueAssign}{is a real value to place for internal and terminal branches (default=0).}

\item{nTimes}{is an integer, number of times to run the branch swapping function.}
}
\value{
A list of three data.frame objects, branch swapping eval for: 1. All brach lengths set to valueAssign (default=1). 2. Only internal braches set to valueAssign (default=1). 3. Only terminal braches set to valueAssign (default=1).
}
\description{
This function tests whether the internal/terminal branch lengths impact the results.
}
\examples{
library(blepd)
data(tree)
data(distribution)
## calculate values

t3 <- evalTerminalvsInternal(tree, distribution, nTimes=10, valueAssign=0.5, branch="terminals")
print.multiBlepd(t3,tabular=FALSE)


}
\author{
Miranda-Esquivel Daniel R.
}
