% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalTree.R
\name{evalTree}
\alias{evalTree}
\title{evalTree}
\usage{
evalTree(tree = tree, distribution = distribution, root = FALSE,
  index = "PD")
}
\arguments{
\item{tree}{is a single tree with T terminals, an APER phylo object.}

\item{distribution}{is a labeled matrix object, with the distribution of T terminals (rows) in A areas (columns).}

\item{root}{is use.root in PD function.}
}
\value{
A data.frame object with 14 fields:  labelTerminal, lowerBranchLength, InitialArea, lowerFinalArea, initialLength, upperBranchLength, upperFinalArea, changeLower, changeUpper, deltaUpper, deltaLower.   deltaPD,  areaDelta, and  abDelta.
}
\description{
The function is a wrap for evalTerminal and calculates whether changes in all the terminal branches length generate changes in the area selected; and when applies, the terminal branch length value for that change(s).
}
\examples{
library(blepd)
data(tree)
data(distribution)
evalTree(tree = tree , distribution = distribution , root = FALSE)


}
\author{
Miranda-Esquivel Daniel R.
}
