% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{findTerminalgivenLength}
\alias{findTerminalgivenLength}
\title{Find terminal branches with lengths exceeding a specified threshold.}
\usage{
findTerminalgivenLength(arbol, valor = 0)
}
\arguments{
\item{arbol}{A phylogenetic tree object in the 'phylo' format.}

\item{valor}{A numeric value representing the length threshold. Branches with lengths greater than this value will be identified. Defaults to 0.0.}
}
\value{
A character vector containing the labels of the terminal branches that meet the length threshold.
}
\description{
This function identifies the terminal branches in a phylogenetic tree that have lengths greater than a given threshold.
}
\examples{

# Find terminal branches with lengths greater than 0.5
long_branches <- findTerminalgivenLength(arbol = tree, valor = 0.5) 

}
