% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{matrix2XY}
\alias{matrix2XY}
\title{Convert a species distribution matrix into a data frame of terminal taxa and their associated distribution.}
\usage{
matrix2XY(distribution = distribution)
}
\arguments{
\item{distribution}{A matrix or data frame representing the presence/absence of species in different sets. Rows represent sets and columns represent species.}
}
\value{
A data frame with two columns: "Terminal" (the species name) and "Area" (the area where the species is present).
}
\description{
This function takes a matrix or data frame representing the presence/absence of species in different sets and converts it into a data frame where each row represents a terminal taxon and its corresponding area.
}
\examples{

# Convert distribution matrix to data frame
xy_df <- matrix2XY(distribution = distribution)

}
