% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printEvalBranch.R
\name{print.evalBranchAll}
\alias{print.evalBranchAll}
\title{print.evalBranchAll}
\usage{
\method{print}{evalBranchAll}(object0)
}
\arguments{
\item{object0}{An `EvalBranch` object or a list containing `EvalBranch` objects.}

\item{compact}{A logical value indicating whether to print the information 
in a compact format (e.g., branch number instead of full branch string) 
(default = TRUE).}
}
\value{
Prints the formatted information to the console.
}
\description{
Prints information from an `EvalBranch` object or a list containing 
`EvalBranch` objects. It summarizes the branch evaluated, initial area, 
modified area after evaluation, and delta value (percentage change).
}
\examples{
library(blepd)  

# Create an EvalBranch object
evalBranch <- EvalBranch(branchToEval = "Branch1", 

# Print information from the object
print.evalBranchAll(evalBranch)

# Create a list of EvalBranch objects
multiEvalResults <- list(evalBranch, evalBranch)

# Print information from the list
print.evalBranchAll(multiEvalResults, compact = FALSE)
}
