% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printEvalBranch.R
\name{printEvalBranch}
\alias{printEvalBranch}
\title{printEvalBranch}
\usage{
printEvalBranch(objectToPrint, compact = TRUE)
}
\arguments{
\item{compact}{Logical indicating a concise format (default = TRUE).}

\item{object}{An `EvalBranch` object or a list containing `EvalBranch` objects.}
}
\value{
Prints the formatted information to the console.
}
\description{
Prints a formatted summary of information from an `EvalBranch` object or a list containing `EvalBranch` objects.
}
\examples{
library(blepd)  

# Create an EvalBranch object
evalBranch <- EvalBranch(branchToEval = "Branch1", 

# Print information from the object
printEvalBranch(evalBranch)

}
\author{
Miranda-Esquivel Daniel R.
}
