% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printSwapBL.R
\name{printSwapBL}
\alias{printSwapBL}
\title{Print Information from a swapBL Object}
\usage{
printSwapBL(objectToPrint, compact = FALSE, tabular = TRUE)
}
\arguments{
\item{objectToPrint}{An `swapBL` object.}

\item{compact}{Logical indicating whether to print a compact output (default: FALSE).}

\item{tabular}{Logical indicating whether to print output in tabular format (default: TRUE).}
}
\value{
Invisibly returns the object.
}
\description{
Prints information from a `swapBL` object. 
The output format can be controlled using the `compact` and `tabular` arguments.
}
\examples{
library(blepd)  

# Create a swapBL object
evalBranch <-  swapBL(tree=tree,
                      distribution = distribution,
                          model = "simpleswap",
                          branch = "terminals",
                          verbose = TRUE
                         )

# Print information from the object
printSwapBL(evalBranch)

}
\author{
Miranda-Esquivel Daniel R.
}
