% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RhinoclemmysData.R
\docType{data}
\name{RhinoclemmysData}
\alias{RhinoclemmysData}
\title{RhinoclemmysData}
\format{
A list with the initial objects, and the resulting analyses: 
RhinoclemmysData$distribution   : A matrix object
RhinoclemmysData$tree           : A phylo object
}
\usage{
RhinoclemmysData
}
\description{
Tree and distribution for _Rhinoclemmys_, an example data analysis
}
\author{
Miranda-Esquivel Daniel R.
}
\keyword{datasets}
