% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swapBL.R
\name{swapBL}
\alias{swapBL}
\title{swapBL: Analyze Effect of Branch Length Swapping on Phylogenetic Diversity}
\usage{
swapBL(
  tree = tree,
  distribution = distribution,
  model = "allswap",
  nTimes = 100,
  root = TRUE,
  index = "PD",
  branch = "terminals",
  percentual = FALSE,
  verbose = TRUE,
  compact = TRUE
)
}
\arguments{
\item{tree}{(required): A phylogenetic tree object in the APER format, containing terminal labels (tips).}

\item{distribution}{(required): A matrix representing the distribution of terminal taxa across sets (rows represent sets, columns represent terminal taxa). Ensure the column names in this matrix match the terminal labels (tips) in the tree.}

\item{model}{(optional, default="allswap"): The type of branch length swapping to perform. Valid options are:
"simpleswap": Swaps the lengths of two randomly chosen branches.
"allswap" (default): Randomly permutes the branch lengths within the specified branch scope (defined by the `branch` argument).
"uniform": Replaces the branch lengths of chosen nodes with random values drawn from a uniform distribution between the minimum and maximum values of the original lengths.}

\item{nTimes}{(optional, default=100): The number of times to repeat the swapping process (iterations).}

\item{root}{(optional, default=TRUE): A logical value indicating whether to use the root node in PD calculations.}

\item{branch}{(optional, default="terminals"): The type of branches to swap lengths for. Valid options are:
"terminals": Swaps lengths among terminal branches.
"internals": Swaps lengths among internal branches.
"all": Swaps lengths among all branches (terminals and internals).}

\item{verbose}{(optional, default=TRUE): A logical value indicating whether to print informative messages during function execution.}

\item{compact}{(optional, default=TRUE): A logical value determining the output content. If `FALSE`, all available information is returned. If `TRUE` (default), a more concise output is provided.}
}
\value{
`initialPD`: The PD value of the original tree.

 `bestInitialArea`: The set(s) with the highest PD in the initial tree.

 `bestModifiedArea`: A data frame summarizing the frequency of each set identified with the highest PD across iterations.

 `model`: The type of branch length swapping performed ("simpleswap", "allswap", or "uniform").

 `nTimes`: The number of times the swapping process was repeated.

 `branch`: The type of branches swapped ("terminals", "internals", or "all").

 `root` (optional): A logical value indicating whether the root node was used during PD calculations (default: TRUE).

 `index` (optional): The PD index used (currently always "PD").
}
\description{
This function evaluates how swapping terminal or internal branch lengths affects Phylogenetic Diversity (PD) in a phylogenetic tree. It achieves this by simulating branch length modifications and assessing the resulting changes in PD.
}
\examples{

library(blepd)

data(tree)

data(distribution)

# Evaluate the effect of swapping all terminal branch lengths 100 times

swapBL(tree = tree, distribution = distribution, model = "allswap", nTimes = 100, branch = "terminals") 

}
\author{
Miranda-Esquivel Daniel R.
}
