% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\docType{data}
\name{tree}
\alias{tree}
\title{A tree with four terminals.}
\format{
An APER object of class `phylo` representing a tree with four terminals.
}
\usage{
tree
}
\description{
An APER object representing a phylogenetic tree with four terminals.
}
\author{
Miranda-Esquivel Daniel R.
}
\keyword{datasets}
