% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\docType{data}
\name{distribution}
\alias{distribution}
\title{Distribution data for four taxa in two areas.}
\format{A matrix object with two areas (rows) and four species (rows).}
\usage{
distribution
}
\description{
A matrix with four species distributions in two areas.
}
\author{
Miranda-Esquivel Daniel R.
}
\keyword{datasets}
