% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalBranch.R
\name{evalBranch}
\alias{evalBranch}
\title{evalBranch}
\usage{
evalBranch(
  tree = tree,
  distribution = distribution,
  branchToEval = branchToEval,
  approach = "lower",
  root = FALSE,
  index = "PD",
  maxMultiplier = 1.01,
  redondeo = 2,
  verbose = FALSE,
  compact = TRUE
)
}
\arguments{
\item{tree}{is a single tree with T terminals, an APER phylo object.}

\item{distribution}{is a labeled matrix object, with the distribution of T terminals (columns) in A areas (rows).}

\item{branchToEval}{is the number/name of the branch to evaluate, "terminals" or "internals" evaluate only those named, while "all", evaluates all terminals/internals.}

\item{approach}{is the type of limit to evaluate, "upper": from the actual length to maxVal [*maxMultiplier], or "lower": from the actual length to 0.0, or "all" (default = "upper").}

\item{root}{is use.root in PD function (default = FALSE).}

\item{maxMultiplier}{is the value to multiply the sum of the branch length values. The upper limit to evaluate will be BL_sum * maxMultiplier (default = 1.01).}

\item{verbose}{is the length of the speech (default = FALSE).}
}
\value{
Returns a S3 object [class blepd] with all the relevant information: whether there is no-change/change in area as we change the terminal branch length, the maxPD difference for the upper/lower limit, the branch length of the change, the best Initial Area, the actual (initial) branch length, and the area selected.
}
\description{
The function calculates whether a change in a branch length generates a change in the area selected; and when applies, the branch length value for that change.
}
\examples{
library(blepd)
data(tree)
data(distribution)
evalBranch(tree = tree , distribution = distribution , 
           branchToEval = "all" ,  approach = "lower" , 
           root = TRUE)


}
\author{
Miranda-Esquivel Daniel R.
}
