% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalBranch.R
\name{evalBranch}
\alias{evalBranch}
\title{evalBranch}
\usage{
evalBranch(
  tree = tree,
  distribution = distribution,
  branchToEval = "terminals",
  approach = "upper",
  root = FALSE,
  index = "PD",
  maxMultiplier = 1.01,
  redondeo = 3,
  verbose = FALSE,
  compact = TRUE,
  printNames = FALSE
)
}
\arguments{
\item{tree}{A single phylogenetic tree in APER format.}

\item{distribution}{A matrix indicating the distribution of terminal taxa across sets.}

\item{branchToEval}{The branch(es) to evaluate (can be a numeric vector of node 
numbers, "terminals", "internals", or "all", default = "terminals").}

\item{approach}{The type of branch length modification: 
"lower" (decrease to zero), "upper" (increase to maximum), 
or "all" for both (default = "upper").}

\item{root}{Logical indicating whether to use the root in PD calculation (default = FALSE).}

\item{index}{The index used for PD calculation (e.g., "PD").}

\item{maxMultiplier}{Multiplier for determining the upper limit of branch length 
modification (BL_sum * maxMultiplier, default = 1.01).}

\item{redondeo}{Number of decimal places for rounding.}

\item{verbose}{Logical indicating whether to print verbose output (default = FALSE).}

\item{compact}{Logical indicating whether to return a compact output (default = TRUE).}

\item{printNames}{Logical indicating whether to print terminal names in output.}
}
\value{
The function returns a list containing information about the effect of branch length modifications on set selection, including:

  `branchToEval`: The evaluated branch(es).

  `bestInitialArea`: The set(s) with the highest PD in the initial tree.

  `bestModifiedArea`: The set(s) with the highest PD in the modified tree.

  `approach`: The type of branch length modification used.

  `delta`: The percentage change in branch length.

  `initialPD`: The initial PD value.

  `modifiedPD`: The PD value after the branch length modification.

  `initialLength`: The initial branch length.

  `finalLength`: The final branch length after modification.

  `areas`: A vector of set names.

  `terminals`: A vector of terminal taxa names.

  `root`: Logical indicating whether the root was used in PD calculation.

  `index`: The PD index used.
}
\description{
This function evaluates the effect of modifying a single branch length on 
Phylogenetic Diversity (PD) and the selection of sets with the highest PD. 
It allows for both decreasing (to zero) and increasing (to a user-defined maximum).
}
\examples{
library(blepd) 
data(tree)
data(distribution)

# Evaluate the effect of increasing the length of terminal branches

result_upper <- evalBranch(tree = tree, distribution = distribution, 
                         branchToEval = "terminals", approach = "upper") 

}
\author{
Miranda-Esquivel Daniel R.
}
