% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalTerminal.R
\name{evalTerminal}
\alias{evalTerminal}
\title{evalTerminal}
\usage{
evalTerminal(tree = tree, distribution = distribution, tipToEval = "taxB",
  approach = "lower", root = FALSE, maxMultiplier = 1.5)
}
\arguments{
\item{tree}{is a single tree with T terminals, an APER phylo object.}

\item{distribution}{is a labeled matrix object, with the distribution of T terminals (rows) in A areas (columns).}

\item{tipToEval}{is the label of the terminal to evaluate.}

\item{approach}{is the type of limit to evaluate, "upper": from the actual length to maxVal, or "lower": from the actual length to 0.}

\item{root}{is use.root in PD function.}

\item{maxMultiplier}{is the value to multiply the sum of the PD values. This value will be the upper limit to evaluate.}
}
\value{
Four fields; and depending on the results -whether there is no-change/change in area as we change the terminal branch length-, the function returns the maxPD+1, the best Initial Area, a dummy value of "*" to indicate there is no change in area and the actual (initial) branch length; or, when there is a change in area selected, the function returns the branch length of the change, the best Initial Area, the area selected, and the actual (initial) branch length.
}
\description{
The function calculates whether a change in the terminal branch length generates a change in area selected; and when applies, the terminal branch length value for that change.
}
\examples{
library(blepd)
data(tree)
data(distribution)
evalTerminal(tree = tree , distribution = distribution , tipToEval = "t1" , 
approach = "lower" , root = FALSE)


}
\author{
Miranda-Esquivel Daniel R.
}
