% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{matrix2XY}
\alias{matrix2XY}
\title{matrix2XY}
\usage{
matrix2XY(distribution = distribution)
}
\arguments{
\item{distribution}{is a labeled matrix object, with the distribution of T terminals (columns) in A areas (rows).}
}
\description{
Converts a distribution object into a XY data.frame, suitable to plot.
}
\examples{
library(blepd)
data(distribution)
matrix2XY(distribution)


}
\author{
Miranda-Esquivel Daniel R.
}
