% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{PDindex}
\alias{PDindex}
\title{@title Calculate Phylogenetic Diversity (PD) or Phylogenetic Endemism (PE) for a given distribution.}
\usage{
PDindex(
  tree = tree,
  distribution = distribution,
  root = FALSE,
  index = "PD",
  percentual = FALSE
)
}
\arguments{
\item{tree}{A phylogenetic tree object in the 'phylo' format.}

\item{distribution}{A matrix or data frame representing the presence/absence of species in different areas. Rows represent areas and columns represent species.}

\item{root}{Logical. If TRUE, the root of the tree is included in the calculations. Defaults to FALSE.}

\item{index}{Character string specifying the index to calculate. Valid options are "PD" for Phylogenetic Diversity and "PE" for Phylogenetic Endemism. Defaults to "PD".}

\item{percentual}{Logical. If TRUE, the results are returned as percentages of the total PD or PE. Defaults to FALSE.}
}
\value{
A numeric vector containing the calculated PD or PE values for each area in the distribution matrix.
}
\description{
This function calculates the Phylogenetic Diversity (PD) or Phylogenetic Endemism (PE) for a given distribution of species across a phylogeny.
}
\examples{

# Calculate PD for each set
PD_values <- PDindex(tree = tree, distribution = distribution) 

}
