% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printEvalBranch.R
\name{printEvalBranch}
\alias{printEvalBranch}
\title{Print Information from an EvalBranch Object}
\usage{
printEvalBranch(objectToPrint, compact = TRUE)
}
\arguments{
\item{compact}{Logical indicating whether to print a concise format (default = TRUE).}

\item{object}{An `EvalBranch` object or a list containing `EvalBranch` objects.}
}
\value{
Invisibly returns the `object`.
}
\description{
Prints a formatted summary of information from an `EvalBranch` object or a list 
containing `EvalBranch` objects.
}
\examples{
library(blepd)  

# Create an EvalBranch object: Evaluate the effect of increasing the length of terminal branches

result_upper <- evalBranch(tree = tree, distribution = distribution, 
                         branchToEval = "terminals", approach = "upper") 

# Print information from the object

printEvalBranch(result_upper)

}
\author{
Miranda-Esquivel Daniel R.
}
