% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printswapBL.r
\name{printswapBL}
\alias{printswapBL}
\title{print.swapBL-Output
This function takes the output of a 'blepd::swapBL' analysis and prints a summary
of the area with the best fit for the dispersal model. It includes the area 
name, frequency of occurrences in that area, and the calculated percentage.}
\usage{
printswapBL(swapBL.output)
}
\arguments{
\item{swapBL.output}{A list containing the output of a 'blepd' analysis, 
typically of class "blepd".}
}
\value{
This function does not return a value, it prints the summary to the console.
}
\description{
print.swapBL-Output
This function takes the output of a 'blepd::swapBL' analysis and prints a summary
of the area with the best fit for the dispersal model. It includes the area 
name, frequency of occurrences in that area, and the calculated percentage.
}
\examples{
# Assuming you have your 'blepd' analysis output in 'my_blepd_output'
print_best_area_summary(my_blepd_output)

}
