% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RhinoclemmysData.R
\docType{data}
\name{RhinoclemmysData}
\alias{RhinoclemmysData}
\title{RhinoclemmysData}
\format{
A list with the following components:
  - `distribution`: A matrix object representing the species distribution.
  - `tree`: A `phylo` object representing the phylogenetic tree.
}
\usage{
RhinoclemmysData
}
\description{
A list containing data for _Rhinoclemmys_, an example dataset for 
phylogenetic diversity analysis.
}
\author{
Miranda-Esquivel Daniel R.
}
\keyword{datasets}
