% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RhinoclemmysData.R
\docType{data}
\name{RhinoclemmysData}
\alias{RhinoclemmysData}
\title{A tree and distribution for Rhinoclemmys}
\format{An APE object, class phylo. A tree dataset with four terminals.}
\usage{
RhinoclemmysData
}
\description{
A tree as an APE object with four terminals.
}
\author{
Miranda-Esquivel Daniel R.
}
\keyword{datasets}
