% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RhinoclemmysData.R
\docType{data}
\name{RhinoclemmysData}
\alias{RhinoclemmysData}
\title{RhinoclemmysData}
\format{
A list with the initial objects, and the resulting analyses: 
RhinoclemmysData$distribution   : a matrix object
RhinoclemmysData$trees          : phylo objects
RhinoclemmysData$tablePD        : a matrix object
RhinoclemmysData$matrixPD       : a matrix object
RhinoclemmysData$nameTrees      : a vector of chars
RhinoclemmysData$nameAreas      : a vector of chars
RhinoclemmysData$nameTerminals  : a vector of chars
RhinoclemmysData$evalPD         : a vector of reals
RhinoclemmysData$sumDeltas      : a vector of reals
RhinoclemmysData$maxPD          : a vector of reals
}
\usage{
RhinoclemmysData
}
\description{
Trees and distributions for _Rhinoclemmys_, an example data analysis
}
\author{
Miranda-Esquivel Daniel R.
}
\keyword{datasets}
