% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swapBL.R
\name{swapBL}
\alias{swapBL}
\title{swapBL: Analyze Effect of Branch Length Swapping on Phylogenetic Diversity}
\usage{
swapBL(
  tree = tree,
  distribution = distribution,
  model = "allswap",
  nTimes = 100,
  root = TRUE,
  index = "PD",
  branch = "terminals",
  percentual = FALSE,
  verbose = TRUE,
  compact = TRUE
)
}
\arguments{
\item{tree}{(required): A phylogenetic tree object in the `ape` format, containing terminal labels (tips).}

\item{distribution}{(required): A matrix representing the distribution of terminal taxa across areas (rows represent areas, columns represent terminal taxa). Ensure the column names in this matrix match the terminal labels (tips) in the tree.}

\item{model}{(optional, default="allswap"): The type of branch length swapping to perform. Valid options are:}

\item{nTimes}{(optional, default=100): The number of times to repeat the swapping process (iterations).}

\item{root}{(optional, default=TRUE): A logical value indicating whether to use the root node in PD calculations.}

\item{branch}{(optional, default="terminals"): The type of branches to swap lengths for. Valid options are:}

\item{verbose}{(optional, default=TRUE): A logical value indicating whether to print informative messages during function execution.}

\item{compact}{(optional, default=TRUE): A logical value determining the output content. If `FALSE`, all available information is returned. If `TRUE` (default), a more concise output is provided.}
}
\description{
swapBL: Analyze Effect of Branch Length Swapping on Phylogenetic Diversity
}
\examples{

library(blepd)

data(tree)

data(distribution)

# Evaluate the effect of swapping all terminal branch lengths 100 times

swapBL(tree = tree, distribution = distribution, model = "allswap", nTimes = 100, branch = "terminals") 

}
\author{
Miranda-Esquivel Daniel R.
}
