% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swapBL.R
\name{swapBL}
\alias{swapBL}
\title{swapBL}
\usage{
swapBL(tree = tree, distribution = distribution, model = "simpleswap",
  nTimes = 100, root = TRUE)
}
\arguments{
\item{tree}{is a single tree with T terminals, an APER phylo object.}

\item{distribution}{is a labeled matrix object, with the distribution of T terminals (rows) in A areas (columns).}

\item{number}{to times the swap is performed.}
}
\value{
Returns a data frame with the frequency of each area recovered.
}
\description{
The function calculates the effect on PD of swapping two terminal brach lengths.
}
\examples{
library(blepd)
data(tree)
data(distribution)
swapBL(tree = tree , distribution = distribution , 
       times = 100)


}
\author{
Miranda-Esquivel Daniel R.
}
