% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\docType{data}
\name{tree}
\alias{tree}
\title{A tree with five terminals.}
\format{An APE object, class phylo. A tree dataset with four terminals.}
\usage{
tree
}
\description{
A tree as an APE object with four terminals.
}
\author{
Miranda-Esquivel Daniel R.
}
\keyword{datasets}
